+++ Keyboards/LanguageKeyboards/Portuguese/PTInterfaceVariables.swift
@@ -101,4 +101,5 @@ func setPTKeyboardLayout() {
   pluralKeyLbl = "Plural"
   pluralPrompt = commandPromptSpacing + "Plural: "
   pluralPromptAndCursor = pluralPrompt + commandCursor
+  isAlreadyPluralMessage = "Já plural"
 }